refactor(filesystem): consolidate semantic projection modules
